+++ models/gpt2/generate.py
@@ -3,8 +3,9 @@ from typing import List, cast
 import torch
 from torch import Tensor
 from torch.nn import functional as F
-from models.gpt2.load import build
 
+from models.helpers import timeit
+from models.gpt2.load import build
 from models.gpt2.tokenizer import Tokenizer
 from models.gpt2.transformer import Transformer, GPTConfig
 
@@ -21,6 +22,7 @@ class GPT2:
     return self
 
   @staticmethod
+  @timeit(desc="Load time", ms=False)
   def from_pretrained(model_desc: str='gpt2'):
     model, tokenizer, config = build(model_desc)
     return GPT2(model, tokenizer, config)
