+++ models/mistral_nonrolling/generate.py
@@ -3,8 +3,36 @@ from typing import List
 import torch
 import torch.nn.functional as F
 
+from models.helpers import timeit
+from models.mistral_nonrolling.config import MistralConfig
 from models.mistral_nonrolling.tokenizer import Tokenizer
 from models.mistral_nonrolling.transformer import Transformer
+from models.mistral_nonrolling.load import build
+
+
+class Mistral:
+  def __init__(self, model: Transformer, tokenizer: Tokenizer, config: MistralConfig):
+    self.model, self.tokenizer, self.config = model, tokenizer, config
+
+  @property
+  def device(self) -> torch.device: return next(self.model.parameters()).device
+  @property
+  def dtype(self) -> torch.dtype: return next(self.model.parameters()).dtype
+
+  def to(self, device: torch.device):
+    self.model = self.model.to(device)
+    return self
+
+  @staticmethod
+  @timeit(desc="Load time", ms=False)
+  def from_pretrained(folder: str, max_seq_len: int, max_batch_size: int, device: torch.device):
+    model, tokenizer, config = build(folder, max_seq_len, max_batch_size)
+    return Mistral(model, tokenizer, config).to(device)
+
+  @torch.no_grad()
+  def generate(self, prompts: List[str], max_tokens: int):
+    return generate(self.model, self.tokenizer, self.device, prompts, max_tokens)
+
 
 @torch.no_grad()
 def generate(model: Transformer, tokenizer: Tokenizer, device: torch.device, prompts: List[str], max_tokens: int):
